Fix signed/unsigned comparisons.

+++ cpp/server/handler.cc
@@ -281,7 +281,7 @@ void HttpHandler::GetEntries(evhttp_request *req) const {
   }
 
   JsonArray json_entries;
-  for (size_t i = start; i <= end; ++i) {
+  for (int i = start; i <= end; ++i) {
     LoggedCertificate cert;
     if (manager_->GetEntry(i, &cert) != CTLogManager::FOUND) {
       return SendError(req, HTTP_BADREQUEST, "Entry not found.");
@@ -351,7 +351,7 @@ void HttpHandler::GetProof(evhttp_request *req) const {
   }
 
   const int tree_size(GetIntParam(query, "tree_size"));
-  if (tree_size < 0 || tree_size > manager_->GetSTH().tree_size()) {
+  if (tree_size < 0 || (uint64_t)tree_size > manager_->GetSTH().tree_size()) {
     return SendError(req, HTTP_BADREQUEST,
                      "Missing or invalid \"tree_size\" parameter.");
   }
